ChainMono, minigame, step, FFButton see descrip
Minigames should now be made using the minigame component.
ChainMono is now being used for every step, further optimization is required, (only update the step that is active is a solution)
Adapted the current chainactions with a StartMaxTime, should make this more readable, a simple bool would suffice.
FFButton is now implemented which is using the IClickable interface .
Removed ignore UI layer from gamemanager prefab.
Updated the recipe controller

MisterWitch scene is also updated with all this functionality. It's now possible to quit and start again in the middle of the mission, it's also possible to replay the minigame after completing the mission.

However we still need to save data that the minigame has been played or the mission has been started. So that when leaving the scene, the game can continue where it left off when re-entering the scene.